trying to  convert func parseResponse  to class
Almost finished converting. Realized that I could place the rest of the func parseResponse into the init of class ParseResponse.
But, there is an issue with the func select. When it's in func parseResponse (175-177), it uses the definition from the swiftSoup package. However, when it's in the class ParseResponse (114-116), it uses the default definition. Once that error is resolved, I can move the track loop (187-190) into the class, and fully remove func parseResponse.